chore(e2e): consolidate test runners into unified e2e command
