+++ e2e/package.json
@@ -3,18 +3,12 @@
   "version": "1.0.0",
   "description": "E2E tests for PocketPal using Appium + WebDriverIO",
   "scripts": {
-    "test:android": "wdio run wdio.android.conf.ts",
-    "test:ios": "wdio run wdio.ios.conf.ts",
-    "test:android:local": "wdio run wdio.android.local.conf.ts",
-    "test:ios:local": "wdio run wdio.ios.local.conf.ts",
-    "test:aws": "npx ts-node scripts/run-aws-device-farm.ts",
-    "test:models:ios": "npx ts-node scripts/run-model-tests.ts --platform ios",
-    "test:models:android": "npx ts-node scripts/run-model-tests.ts --platform android",
-    "crash-repro": "npx ts-node scripts/run-crash-repro.ts",
-    "crash-repro:local": "npx ts-node scripts/run-crash-repro.ts --local",
-    "pipeline": "npx ts-node scripts/run-e2e-pipeline.ts",
-    "pipeline:ios": "npx ts-node scripts/run-e2e-pipeline.ts --platform ios",
-    "pipeline:android": "npx ts-node scripts/run-e2e-pipeline.ts --platform android",
+    "e2e": "npx ts-node scripts/run-e2e.ts",
+    "e2e:ios": "npx ts-node scripts/run-e2e.ts --platform ios",
+    "e2e:android": "npx ts-node scripts/run-e2e.ts --platform android",
+    "e2e:models:ios": "npx ts-node scripts/run-e2e.ts --platform ios --each-model",
+    "e2e:models:android": "npx ts-node scripts/run-e2e.ts --platform android --each-model",
+    "e2e:aws": "npx ts-node scripts/run-aws-device-farm.ts",
     "typecheck": "tsc --noEmit"
   },
   "packageManager": "yarn@1.22.22",
