+++ e2e/wdio.ios.local.conf.ts
@@ -11,6 +11,8 @@
  *   E2E_DEVICE_UDID       - Device UDID (default: undefined = simulator auto-selection)
  *   E2E_APP_PATH           - Path to .app or .ipa (default: release simulator build)
  *   E2E_APPIUM_PORT        - Appium server port (default: 4723)
+ *   E2E_XCODE_ORG_ID      - Apple Development Team ID (required for real devices)
+ *   E2E_XCODE_SIGNING_ID  - Code signing identity (default: 'Apple Development')
  */
 
 import {config as sharedConfig} from './wdio.shared.conf';
@@ -22,6 +24,8 @@ const PLATFORM_VERSION = process.env.E2E_PLATFORM_VERSION || '26.0';
 const DEVICE_UDID = process.env.E2E_DEVICE_UDID; // undefined = simulator auto-selection
 const APP_PATH = process.env.E2E_APP_PATH || '../ios/build/Build/Products/Release-iphonesimulator/PocketPal.app';
 const APPIUM_PORT = parseInt(process.env.E2E_APPIUM_PORT || '4723', 10);
+const XCODE_ORG_ID = process.env.E2E_XCODE_ORG_ID;
+const XCODE_SIGNING_ID = process.env.E2E_XCODE_SIGNING_ID || 'Apple Development';
 
 export const config: Options.Testrunner = {
   ...sharedConfig,
@@ -43,6 +47,11 @@ export const config: Options.Testrunner = {
       'appium:autoAcceptAlerts': true,
       // Only include UDID if explicitly set (real devices need it)
       ...(DEVICE_UDID && {'appium:udid': DEVICE_UDID}),
+      // Code signing for real devices (WebDriverAgent needs to be signed)
+      ...(XCODE_ORG_ID && {
+        'appium:xcodeOrgId': XCODE_ORG_ID,
+        'appium:xcodeSigningId': XCODE_SIGNING_ID,
+      }),
     },
   ],
